suggesting replacing types_handler type detection with simpler implementation.
