+++ src/shmessy/types/datetime_.py
@@ -4,9 +4,9 @@ import numpy as np
 from numpy import ndarray
 from pandas import Series, to_datetime
 
-from ..schema import InferredField
-from . import validate_strptime_pattern
 from .base import BaseType
+from .utils import validate_strptime_pattern
+from ..schema import InferredField
 
 
 class DatetimeType(BaseType):
